feat(pipeline): add support for custom runner tags in GithubCDKPipeline (#59)
This commit introduces the ability to specify custom runner tags for the GithubCDKPipeline. The GithubCDKPipelineOptions interface now accepts an optional `runnerTags` array that designates the runners on which the pipeline jobs (synth, assetUpload, and deploy) should run. If no `runnerTags` are provided, it defaults to 'ubuntu-latest'. This enhancement gives users the flexibility to use self-hosted or specialized runners for their CI/CD pipelines.