feat(pipelines): update pkgNamespace to be optional and enhance validation (#141)
The `pkgNamespace` property in the CDK pipeline options has been updated
to be optional. This change allows for more flexible configurations,
particularly when versioned artifacts are not required. Additionally,
validation has been added to ensure that `pkgNamespace` is provided when
manual approvals are used in the pipeline stages. Documentation has been
updated to reflect these changes, and tests have been added to verify
the new behavior.

Fixes #