chore: update dependencies and GitHub Actions configurations
This commit updates the `jsii`, `jsii-diff`, `jsii-pacmak`, `jsii-rosetta`, and `projen` dependencies to their latest versions. It also modifies GitHub Actions workflows to use the new bot identity for commits and updates the Node.js version in the workflows. Additionally, it enhances the `eslint` task environment and updates the npm-check-updates command to a newer version.

Changes include:
- Updated `jsii` to `~5.9`, `jsii-diff` to `^1.124.0`, `jsii-pacmak` to `^1.124.0`, `jsii-rosetta` to `~5.9`, and `projen` to `0.98.12`.
- Adjusted GitHub Actions user identity to `github-actions[bot]`.
- Set Node.js version to `24.x` in workflows.
- Added `NODE_NO_WARNINGS` environment variable to the `eslint` task.
- Updated npm-check-updates command to version `18`.

These changes ensure compatibility with the latest versions and improve the CI/CD process.